Add files via upload.
These contain scripts that instantiate the environment for the openai framework. The training script which uses the q learning algorithm is provided and parameters are modified in the params file under util. Finally, modifications have been made to the file such as python3, print statements needed brackets etc since the file was written for older versions of Python. Plotting rewards script needs to be updated.